Make tests compatible with PHPUnit 5 API
Updates the usage of the mock API. PHPUnit 5 is the current stable
release so upgrading this will ensure a bigger life time to the
test suite. The constraint on `composer.json` still uses 4.8 as version
as it is still compatible with PHP 5.3, the intention is to tag a 1.0
version of this package with PHP 5.3 compatibility to later drop it.

Creates scripts to ease tests during development and continuous
integration. Bootstrapping can be done completely via Composer and
PHPUnit configuration.

Configures a [build matrix][1] for Travis:

* Favor an environment variable to disable TLS.
* Mark non maintained PHP versions as allowed failures.
* Favor composer scripts instead of commands.
* Only send coverage after a success.

There are still problems needing fixes regarding the test suite, they
need to be addressed separately as they will cause bigger conflicts while
merging this code with other (pending) contributions:

* Some test cases without the suffix `Test.php` are not being executed.
* Avoid exclusion of functional tests.
* Separate test cases directory from concrete classes being used.

[1]: https://docs.travis-ci.com/user/customizing-the-build/#Build-Matrix