Removes functional test suite
No tests in this suite were executed, anyway. We can always recover them
from the version control system, but maybe we will never do that since
unit tests are way better.

In order to have functional tests, one would still need to solve the
issue of authentication. Providing a token for continuous integration
would expose that token to anyone, which is not cool even if the user is
just intended for that.